[DEV-1362] Fixed build
