+++ src/subdomains/generic/user/models/user-data/user-data.service.ts
@@ -27,14 +27,12 @@ import { WebhookService } from '../../services/webhook/webhook.service';
 import { AccountType } from './account-type.enum';
 import { KycUserDataDto } from '../kyc/dto/kyc-user-data.dto';
 import { LinkService } from '../link/link.service';
-import { UserService } from '../user/user.service';
 
 @Injectable()
 export class UserDataService {
   constructor(
     private readonly userDataRepo: UserDataRepository,
     private readonly userRepo: UserRepository,
-    private readonly userService: UserService,
     private readonly bankDataRepo: BankDataRepository,
     private readonly countryService: CountryService,
     private readonly languageService: LanguageService,
@@ -307,8 +305,10 @@ export class UserDataService {
 
     // activate users
     if (master.hasActiveUser) {
+      await this.userDataRepo.activateUserData(master);
+
       for (const user of master.users) {
-        await this.userService.activateUser(user);
+        await this.userRepo.activateUser(user);
       }
     }
   }
